requirements.txt
# inScanLan - Network Scanner Requirements
# Python 3.7+

# Note: tkinter è incluso nella distribuzione standard di Python per Windows
# Non sono necessarie dipendenze esterne per la funzionalità base

# Per la compilazione in EXE:
pyinstaller>=5.0

# Per la conversione del logo JPG in ICO:
pillow>=9.0

# Opzionali (per funzionalità avanzate):
# scapy>=2.5.0  # Per scansione avanzata di rete (richiede Npcap su Windows)
# python-nmap>=0.7.1  # Per integrazione con Nmap (richiede Nmap installato)
